requirements.txt
fastapi
uvicorn
psycopg2-binary
langchain
langchain-openai
pgvector
sqlalchemy
python-dotenv
